fix dew point calcul

+++ WeatherStation/Weather.cpp
@@ -13,16 +13,15 @@
 Weather::Weather() {
 }
 
-float Weather::dew_point(float T, float Rh) {  
+double Weather::dew_point(float T, float Rh) {  
   /*
    * T  : temperature in celsius (C)
    * Rh : Relative Humidity in percent (%)
    * B  : Intermediate value (no units)
    * D  : Dewpoint in Centigrade (C) degrees
    */
-
-  float B = (log(Rh / 100) + ((17.27 * T) / (237.3 + T))) / 17.27;
-  float D = (237.3 * B) / (1 - B);
+  double B = (log(Rh / 100) + ((17.27 * T) / (237.3 + T))) / 17.27;
+  double D = (237.3 * B) / (1 - B);
   return D;
 }
